Renamed the files and split the content as per PR review
