+++ 06_puppy/jimbotech/mapstore.go
@@ -7,29 +7,8 @@ import (
 	"github.com/google/uuid"
 )
 
-// Storer defines standard CRUD operations for Puppy
-type Storer interface {
-	CreatePuppy(p *Puppy) (uint32, error)
-	ReadPuppy(ID uint32) (*Puppy, error)
-	UpdatePuppy(ID uint32, Puppy *Puppy) error
-	DeletePuppy(ID uint32) error
-}
-
-// Puppy stores puppy details.
-type Puppy struct {
-	ID     uint32
-	Breed  string
-	Colour string
-	Value  string
-}
-
 // MapStore stores puppies.
 type MapStore map[uint32]*Puppy
-
-// mapTest used during testing to verify underlaying map changes
-type mapTest interface {
-	length() int
-}
 
 // length used for testing.
 func (s MapStore) length() int {
