+++ 06_puppy/jimbotech/storer_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -32,22 +31,6 @@ func (s *storesSuite) SetupTest() {
 		s.Fail("Unknown Storer implementation")
 	}
 	s.mapper = s.store.(mapTest)
-}
-
-// TestMapStoreWithoutContructor may not be required as it may not be
-// possible to create an instance of that type without using the constructor
-func TestMapStoreWithoutContructor(t *testing.T) {
-	var puppyStore MapStore
-	pup := Puppy{1, "kelpie", "brown", "indispensable"}
-
-	_, err := puppyStore.CreatePuppy(&pup)
-	assert.Equal(t, ErrNotConstructed, err)
-	err = puppyStore.UpdatePuppy(1, &pup)
-	assert.Equal(t, ErrNotConstructed, err)
-	_, err = puppyStore.ReadPuppy(1)
-	assert.Equal(t, ErrNotConstructed, err)
-	err = puppyStore.DeletePuppy(1)
-	assert.Equal(t, ErrNotConstructed, err)
 }
 
 // TestCreateSuccess add to the store and verify
